Added endpoints to the notebooksrouter

+++ routers/notebooks.js
@@ -45,7 +45,7 @@ router.delete(
 
 //get all notebooks of all users
 router.get("/", async (req, res) => {
-  const userId = req.user.id;
+  const { id: userId } = req.user;
   try {
     const userNotebooks = await Notebook.findAll();
     res.status(200).send(userNotebooks);
@@ -79,7 +79,7 @@ router.get("/:notebookId", authMiddleware, async (req, res) => {
 
 //Create a new notebook
 router.post("/", authMiddleware, async (req, res) => {
-  const { userId } = req.user;
+  const { id: userId } = req.user;
   const { name, subjectId } = req.body;
   if ((!name, !subjectId)) {
     res.status(404).send({ message: "Please fill in all input fields" });
@@ -96,14 +96,54 @@ router.post("/", authMiddleware, async (req, res) => {
   }
 });
 
+//Delete a notebook
+router.delete("/:notebookId", authMiddleware, async (req, res, next) => {
+  const { id: userId } = req.user;
+  const notebookId = parseInt(req.params.notebookId);
+
+  if (!notebookId) {
+    res.status(404).send({ message: "Select a notebook to delete" });
+  }
+  try {
+    const notebookToDelete = await Notebook.findByPk(notebookId);
+
+    if (!notebookToDelete) {
+      return res.status(400).send({ message: "Notebook does not exist" });
+    }
+    if (notebookToDelete.userId !== userId) {
+      return res.status(400).send({ message: "you're not authorized" });
+    }
+
+    console.log("THIS IS DELETE NOTEBOOK", notebookId);
+
+    const deletedNotebook = await Notebook.destroy({
+      where: { id: notebookId },
+    });
+
+    return res.status(200).json(deletedNotebook);
+  } catch (e) {
+    next(e);
+  }
+});
+
 //Create a new note
-router.post("/:notebookId/notes", async (req, res) => {
-  const { notebookId } = req.params;
+router.post("/:notebookId/notes", authMiddleware, async (req, res) => {
+  const notebookArray = req.user.dataValues.notebooks;
+  const notebookId = parseInt(req.params.notebookId);
   const { title, content, imageUrl, typeOfNote } = req.body;
   if ((!title, !content, !imageUrl, !typeOfNote)) {
     res.status(404).send({ message: "Please fill in all input fields" });
   }
   try {
+    const notebookIdArray = notebookArray.map(
+      (notebook) => notebook.dataValues.id
+    );
+
+    console.log("THIS IS IDARRAY", notebookIdArray, notebookId);
+    if (!notebookIdArray.includes(notebookId)) {
+      return res.status(400).send({ message: "you're not authorized" });
+    }
+
     const newNote = await Note.create({
       notebookId,
       title,
@@ -119,33 +159,53 @@ router.post("/:notebookId/notes", async (req, res) => {
 
 //Delete a note from notebook
 router.delete("/:notebookId/notes", authMiddleware, async (req, res) => {
-  const { notebookId } = req.params;
+  const notebookId = parseInt(req.params.notebookId);
   const { noteId } = req.body;
+  const notebookArray = req.user.dataValues.notebooks;
+
   if (!noteId) {
     res.status(404).send({ message: "Please select a note to delete" });
   }
+
   try {
-    const newNote = await Note.destroy({
+    const notebookIdArray = notebookArray.map(
+      (notebook) => notebook.dataValues.id
+    );
+
+    if (!notebookIdArray.includes(notebookId)) {
+      return res.status(400).send({ message: "you're not authorized" });
+    }
+
+    const deletedNote = await Note.destroy({
       where: [{ id: noteId }, { notebookId }],
     });
-    res.status(200).json(newNote);
+    res.status(200).json(deletedNote);
   } catch (e) {
     next(e);
   }
 });
 
-//Edit a note from notebook
-router.patch("/:notebookId/notes", authMiddleware, async (req, res) => {
-  const { notebookId } = req.params;
+// Edit a note from notebook
+router.patch("/:notebookId/notes/:noteId", authMiddleware, async (req, res) => {
+  const notebookArray = req.user.dataValues.notebooks;
+  const notebookId = parseInt(req.params.notebookId);
+  const noteId = parseInt(req.params.noteId);
   const { title, content, imageUrl, typeOfNote } = req.body;
   if (!title && !content && !imageUrl && !typeOfNote) {
-    res.status(404).send({ message: "Please select a note to delete" });
+    res.status(404).send({ message: "Please fill in all input fields" });
   }
   try {
-    const newNote = await Note.destroy({
-      where: [{ id: noteId }, { notebookId }],
-    });
-    res.status(200).json(newNote);
+    const notebookIdArray = notebookArray.map(
+      (notebook) => notebook.dataValues.id
+    );
+
+    if (!notebookIdArray.includes(notebookId)) {
+      return res.status(400).send({ message: "you're not authorized" });
+    }
+
+    const noteToUpdate = await Note.findByPk(noteId);
+    const edittedNote = await noteToUpdate.update(req.body);
+    return res.status(200).send(edittedNote);
   } catch (e) {
     next(e);
   }
